Rename api_v2/ to api/ (#347)
* Remove unused specific api v1 exceptions

* Remove v1 ApiError

* Rename api_v2 to api

* Fix access to old api exception field

* Update setup.cfg and setup.py

* Update CHANGELOG section header

* Fix merge
