+++ qiskit/providers/ibmq/api/exceptions.py
@@ -12,53 +12,50 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Exceptions for IBMQ Connector."""
+"""Exceptions related to the IBM Q Experience API."""
+
+from typing import Any
+from requests.exceptions import RequestException
 
 from ..exceptions import IBMQError
 
 
 class ApiError(IBMQError):
-    """IBMQConnector API error handling base class."""
-
-    def __init__(self, usr_msg=None, dev_msg=None):
-        """ApiError.
-
-        Args:
-            usr_msg (str): Short user facing message describing error.
-            dev_msg (str or None): More detailed message to assist
-                developer with resolving issue.
-        """
-        super().__init__(usr_msg)
-        self.usr_msg = usr_msg
-        self.dev_msg = dev_msg
-
-    def __repr__(self):
-        return repr(self.dev_msg)
-
-    def __str__(self):
-        return str(self.usr_msg)
-
-
-class BadBackendError(ApiError):
-    """Unavailable backend error."""
-
-    def __init__(self, backend):
-        """BadBackendError.
-
-        Args:
-            backend (str): name of backend.
-        """
-        usr_msg = 'Could not find backend "{0}" available.'.format(backend)
-        dev_msg = ('Backend "{0}" does not exist. Please use '
-                   'available_backends to see options').format(backend)
-        super().__init__(usr_msg, dev_msg)
-
-
-class CredentialsError(ApiError):
-    """Exception associated with bad server credentials."""
+    """Generic IBM Q API error."""
     pass
 
 
-class RegisterSizeError(ApiError):
-    """Exception due to exceeding the maximum number of allowed qubits."""
+class RequestsApiError(ApiError):
+    """Exception re-raising a RequestException."""
+    def __init__(
+            self,
+            original_exception: RequestException,
+            *args: Any
+    ) -> None:
+        self.original_exception = original_exception
+        super().__init__(*args)
+
+
+class WebsocketError(ApiError):
+    """Exceptions related to websockets."""
+    pass
+
+
+class WebsocketIBMQProtocolError(WebsocketError):
+    """Exceptions related to IBM Q protocol error."""
+    pass
+
+
+class WebsocketAuthenticationError(WebsocketError):
+    """Exception caused during websocket authentication."""
+    pass
+
+
+class WebsocketTimeoutError(WebsocketError):
+    """Timeout during websocket communication."""
+    pass
+
+
+class AuthenticationLicenseError(ApiError):
+    """Exception due to user not accepting latest license agreement via web."""
     pass
