+++ qiskit/providers/ibmq/api/rest/auth.py
