+++ setup.cfg
@@ -11,6 +11,3 @@ strict_equality = True
 disallow_untyped_calls = True
 disallow_untyped_defs = True
 disallow_incomplete_defs = True
-
-[mypy-qiskit.providers.ibmq.api.*]
-ignore_errors = True
