+++ setup.py
@@ -64,9 +64,8 @@ setup(
     keywords="qiskit sdk quantum api ibmq",
     packages=['qiskit.providers.ibmq',
               'qiskit.providers.ibmq.api',
-              'qiskit.providers.ibmq.api_v2',
-              'qiskit.providers.ibmq.api_v2.clients',
-              'qiskit.providers.ibmq.api_v2.rest',
+              'qiskit.providers.ibmq.api.clients',
+              'qiskit.providers.ibmq.api.rest',
               'qiskit.providers.ibmq.circuits',
               'qiskit.providers.ibmq.credentials',
               'qiskit.providers.ibmq.job',
